refactor: move the zip utilities from data/ into util/.
* fits better there.
