+++ lib/data/attachments.js
@@ -1,5 +1,5 @@
 const { join } = require('path');
-const { zipPart } = require('./zip');
+const { zipPart } = require('../util/zip');
 const sanitize = require('sanitize-filename');
 
 const streamAttachments = (inStream) => {
