+++ lib/data/csv.js
@@ -5,7 +5,7 @@ const { toTraversable, traverseFirstChild } = require('../util/xml');
 const { convertToJson } = require('fast-xml-parser');
 const { incr, get, ensureArray } = require('../util/util');
 const csv = require('csv-stringify');
-const { zipPart } = require('./zip');
+const { zipPart } = require('../util/zip');
 const { getFormSchema, flattenSchemaStructures } = require('./schema');
 
 const _requiredIdFields = (idFields, fields) => {
