+++ lib/resources/submissions.js
@@ -8,7 +8,7 @@ const Option = require('../util/option');
 const Problem = require('../problem');
 const { streamJoinedCsvs } = require('../data/csv');
 const { streamAttachments } = require('../data/attachments');
-const { zipStreamFromParts } = require('../data/zip');
+const { zipStreamFromParts } = require('../util/zip');
 
 // multipart things:
 const multer = require('multer');
